+++ src/collective/contact/membrane/behaviors/personmembraneuser.py
@@ -114,8 +114,10 @@ class PersonMembraneUserGroups(grok.Adapter):
         person = self.context
         groups = {}
         for held_position in person.get_held_positions():
-            group = IGroup(held_position.position.to_object)
-            group_id = group.getGroupId()
-            groups[group_id] = 1
+            position = held_position.position
+            if position:
+                group = IGroup(position.to_object)
+                group_id = group.getGroupId()
+                groups[group_id] = 1
 
         return tuple(groups.keys())