+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog
 1.0.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Fix getGroupsForPrincipal so it does not crash if no position
+  [ebrehault]
 
 
 1.0.1 (2014-09-15)